# frontend/control-tower/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentRunView": {
        "additionalProperties": false,
        "properties": {
          "agent": {
            "title": "Agent",
            "type": "string"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "cost_usd": {
            "default": 0.0,
            "title": "Cost Usd",
            "type": "number"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "input_tokens": {
            "default": 0,
            "title": "Input Tokens",
            "type": "integer"
          },
          "llm_calls": {
            "default": 0,
            "title": "Llm Calls",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "output_tokens": {
            "default": 0,
            "title": "Output Tokens",
            "type": "integer"
          },
          "po_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Po Id"
          },
          "po_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Po Name"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "trace_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Url"
          }
        },
        "required": [
          "run_id",
          "agent",
          "status"
        ],
        "title": "AgentRunView",
        "type": "object"
      },
...
      "DemandDay": {
        "additionalProperties": false,
        "properties": {
          "day": {
            "format": "date",
            "title": "Day",
            "type": "string"
          },
          "delivered": {
            "title": "Delivered",
            "type": "number"
          },
          "ordered": {
            "title": "Ordered",
            "type": "number"
          }
        },
        "required": [
          "day",
          "ordered",
          "delivered"
        ],
        "title": "DemandDay",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LineDemand": {
        "additionalProperties": false,
        "description": "Daily demand behind one line, with the forecast the line used (for the sparkline).",
        "properties": {
          "days": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DemandDay"
            },
            "title": "Days",
            "type": "array"
          },
          "forecast_daily": {
            "title": "Forecast Daily",
            "type": "number"
          },
          "line_id": {
            "title": "Line Id",
            "type": "string"
          },
          "product_id": {
            "title": "Product Id",
            "type": "integer"
          },
          "sigma_daily": {
            "title": "Sigma Daily",
            "type": "number"
          },
          "since": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since"
          },
          "until": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          }
        },
        "required": [
          "line_id",
          "product_id",
          "forecast_daily",
          "sigma_daily"
        ],
        "title": "LineDemand",
        "type": "object"
      },
...
      "PlanningLineRow": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted"
          },
          "applied": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied"
          },
          "applied_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "line": {
            "$ref": "#/components/schemas/ReplenishmentLine"
          }
        },
        "required": [
          "line"
        ],
        "title": "PlanningLineRow",
        "type": "object"
      },
      "PlanningOverrides": {
        "additionalProperties": false,
        "description": "Parameters a planner may try in a ``what_if`` (no writes) or force in a review.",
        "properties": {
          "lead_time_days": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Time Days"
          },
          "max_coverage_days": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Coverage Days"
          },
          "review_period_days": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Period Days"
          },
          "service_level": {
            "anyOf": [
              {
                "exclusiveMaximum": 1.0,
                "exclusiveMinimum": 0.5,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Level"
          }
        },
        "title": "PlanningOverrides",
        "type": "object"
      },
...
      "PlanningRunDetail": {
        "additionalProperties": false,
        "properties": {
          "lines": {
            "items": {
              "$ref": "#/components/schemas/PlanningLineRow"
            },
            "title": "Lines",
            "type": "array"
          },
          "run": {
            "$ref": "#/components/schemas/PlanningRunRow"
          }
        },
        "required": [
          "run",
          "lines"
        ],
        "title": "PlanningRunDetail",
        "type": "object"
      },
      "PlanningRunRow": {
        "additionalProperties": false,
        "properties": {
          "approval_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "totals": {
            "additionalProperties": {
              "type": "number"
            },
            "default": {},
            "title": "Totals",
            "type": "object"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "warehouse_id": {
            "title": "Warehouse Id",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "case_id",
          "kind",
          "as_of",
          "warehouse_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "PlanningRunRow",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WhatIfRequest": {
        "additionalProperties": false,
        "properties": {
          "line_id": {
            "title": "Line Id",
            "type": "string"
          },
          "overrides": {
            "$ref": "#/components/schemas/PlanningOverrides"
          }
        },
        "required": [
          "line_id",
          "overrides"
        ],
        "title": "WhatIfRequest",
        "type": "object"
      },
      "WhatIfResponse": {
        "additionalProperties": false,
        "properties": {
          "baseline": {
            "$ref": "#/components/schemas/ReplenishmentLine"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "simulated": {
            "$ref": "#/components/schemas/ReplenishmentLine"
          }
        },
        "required": [
          "baseline",
          "simulated",
          "run_id"
        ],
        "title": "WhatIfResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "director",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/planning/runs": {
      "get": {
        "operationId": "list_planning_runs_api_planning_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanningRunRow"
                  },
                  "title": "Response List Planning Runs Api Planning Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Planning Runs",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/planning/runs/{run_id}": {
      "get": {
        "operationId": "get_planning_run_api_planning_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningRunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Planning Run",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/planning/runs/{run_id}/lines/{line_id}/demand": {
      "get": {
        "description": "The last ``days`` of demand for the line's product, from the planner.",
        "operationId": "line_demand_api_planning_runs__run_id__lines__line_id__demand_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "line_id",
            "required": true,
            "schema": {
              "title": "Line Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 90,
              "maximum": 730,
              "minimum": 7,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineDemand"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Line Demand",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/planning/runs/{run_id}/what-if": {
      "post": {
        "description": "Recompute one line with other parameters; nothing is written anywhere.",
        "operationId": "what_if_api_planning_runs__run_id__what_if_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatIfRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhatIfResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "What If",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/runs": {
      "get": {
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "agent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentRunView"
                  },
                  "title": "Response List Runs Api Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}